// src/app/shared/models/activity-weight.ts
export interface ActivityWeight {
  highEndCarbPerGram: number;
  highEndFatPerGram: number;
  highEndMeatPerDay: number;
  highEndProteinPerGram: number;
  highEndTotalCaloriesTotal: number;
  lowEndCarbPerGram: number;
  lowEndFatPerGram: number;
  lowEndMeatPerDay: number;
  lowEndProteinPerGram: number;
  lowEndTotalCaloriesTotal: number;
}